+++ src/engine/session/utils/settings.ts
@@ -9,7 +9,7 @@ export const getDefaultSettings = () => ({
   hide_sensitive: true,
   report_analytics: true,
   auto_authenticate: false,
-  min_wot_score: 0,
+  min_wot_score: 1,
   imgproxy_url: env.get().IMGPROXY_URL,
   dufflepud_url: env.get().DUFFLEPUD_URL,
   multiplextr_url: env.get().MULTIPLEXTR_URL,